requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
PyMuPDF
requests
openai